Fix loading the same metadata for all videos
