+++ docker/.env.docker
@@ -1,6 +1,7 @@
 APP_SCHEME=http
 APP_HOST=app.test
 APP_PORT=8080
+REVERB_PORT=8080
 
 DB_HOST=db
 DB_PORT=5432
